ci: the Rust gate is green on every leg — fmt on tm-walk's tests, tv_usec widened through i128
CI has failed at "Check the native core" on all four legs since 0bc3aa4,
before a single test ran, for two reasons that never showed on the host:

1. `cargo fmt --all -- --check` rejected two hunks in
   crates/tm-walk/tests/walk.rs — the paused-governor cancel test was added
   without running fmt on that crate. `cargo fmt -p tm-walk`; only that file
   changed.

2. `cargo clippy -D warnings` on the Linux runners rejected
   `i64::from(tv.tv_usec)` in tm-governor's rusage sampler as
   `useless_conversion`: `tv_usec` is already `i64` on linux-gnu (it is
   `i32` on macOS, which is why the host never saw it). A plain `as f64`
   would have traded that for `cast_lossless` on macOS, where the cast is
   lossless and pedantic is on. Widening through `i128` is a real
   conversion on both, and f64 has no `From<i128>`, so both targets lint
   clean without a suppression. The computed value is unchanged.

Verified as CI's step runs it: `cargo fmt --all -- --check`, `cargo clippy
--workspace --all-targets -- -D warnings`, `cargo test --workspace` (126
passed, 3 ignored) on the host, and the same clippy line against
x86_64-unknown-linux-gnu and x86_64-pc-windows-msvc; then the module
rebuilt and `npm test` 2,758 tests, 2,752 passed, 0 failed, 6 skipped;
`npm run typecheck` and `node scripts/build-ui.js --check` clean.

+++ native/treemap-core/crates/tm-governor/src/sample.rs
@@ -217,9 +217,12 @@ fn rusage_cpu_seconds() -> Option<f64> {
     if rc != 0 {
         return None;
     }
-    // tv_usec is i32 on macOS and i64 on Linux; i64::from takes both.
+    // tv_usec is i32 on macOS and i64 on Linux. Widening through i128 is a
+    // real conversion on both (so clippy's useless_conversion stays quiet on
+    // Linux), and f64 has no From<i128>, so the final cast trips no
+    // cast_lossless on macOS either.
     let seconds = |tv: libc::timeval| {
-        tv.tv_sec as f64 + i64::from(tv.tv_usec) as f64 / MICROSECONDS_PER_SECOND
+        tv.tv_sec as f64 + i128::from(tv.tv_usec) as f64 / MICROSECONDS_PER_SECOND
     };
     Some(seconds(usage.ru_utime) + seconds(usage.ru_stime))
 }
